requirements.txt
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
qdrant-client==1.9.2
streamlit>=1.29.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pypdf>=4.0.0